Remove debugging output and update license.
