+++ nginx_module/src/ngx_http_latency_recorder.c
@@ -1,3 +1,29 @@
+// Copyright (C) 2002-2016 Igor Sysoev
+// Copyright (C) 2011-2016 Nginx, Inc.
+// Copyright (C) 2020 Google Inc.
+// All rights reserved.
+//
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided that the following conditions
+// are met:
+// 1. Redistributions of source code must retain the above copyright
+//    notice, this list of conditions and the following disclaimer.
+// 2. Redistributions in binary form must reproduce the above copyright
+//    notice, this list of conditions and the following disclaimer in the
+//    documentation and/or other materials provided with the distribution.
+//
+// THIS SOFTWARE IS PROVIDED BY THE AUTHOR AND CONTRIBUTORS ``AS IS'' AND
+// ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+// ARE DISCLAIMED.  IN NO EVENT SHALL THE AUTHOR OR CONTRIBUTORS BE LIABLE
+// FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+// DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS
+// OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION)
+// HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+// LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY
+// OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
+// SUCH DAMAGE.
+////////////////////////////////////////////////////////////////////////////////
 
 #include "ngx_http_latency_recorder.h"
 #include "ngx_http_latency_storage.h"
@@ -17,19 +43,15 @@ void *ngx_http_latency_create_loc_conf(ngx_conf_t *cf)
   }
 
   conf->enable = NGX_CONF_UNSET;
-  ngx_log_stderr(0, "reached latency filter create");
-
   return conf;
 }
 
 char *ngx_http_latency_merge_loc_conf(ngx_conf_t *cf, void *parent, void *child)
 {
-  ngx_log_stderr(0, "reached latency filter merge");
   ngx_http_latency_conf_t *prev = parent;
   ngx_http_latency_conf_t *conf = child;
 
   ngx_conf_merge_value(conf->enable, prev->enable, 0);
-
   return NGX_CONF_OK;
 }
 
@@ -38,15 +60,10 @@ ngx_int_t ngx_http_latency_init(ngx_conf_t *cf)
   ngx_http_core_main_conf_t  *core_main_config;
   ngx_http_handler_pt *handler;
 
-  ngx_log_stderr(0, "reached latency filter init");
-  //ngx_http_next_header_filter = ngx_http_top_header_filter;
-  //ngx_http_top_header_filter = ngx_http_latency_header_filter;
-
   core_main_config = ngx_http_conf_get_module_main_conf(cf, ngx_http_core_module);
   handler = ngx_array_push(&core_main_config->phases[NGX_HTTP_LOG_PHASE].handlers);
   *handler = ngx_http_record_latency;
 
-  ngx_log_stderr(0, "reached end of latency filter init");
   return NGX_OK;
 }
 
@@ -58,29 +75,21 @@ static ngx_int_t ngx_http_record_latency(ngx_http_request_t *r)
   ngx_http_latency_main_conf_t *main_conf;
   ngx_http_latency_shm_t *latency_record;
 
-  ngx_log_stderr(0, "reached latency filter");
   conf = ngx_http_get_module_loc_conf(r, ngx_http_latency_stub_status_module);
 
   if (!conf->enable) {
     return NGX_DECLINED;
-    //return ngx_http_next_header_filter(r);
   }
 
   main_conf = ngx_http_get_module_main_conf(r, ngx_http_latency_stub_status_module);
   if (main_conf == NULL) {
     ngx_log_stderr(0, "error: no main config");
   }
-  if (main_conf->latency_bucket_bounds == NULL) {
-    ngx_log_stderr(0, "error no latency bounds");
-  }
-  ngx_log_stderr(0, "reached latency filter after enabled check");
-  ngx_log_stderr(0, "max_exponent %d", main_conf->max_exponent);
 
   ngx_time_update();
   now = ngx_timeofday();
   request_time = (ngx_int_t)((now->sec - r->start_sec) * 1000 + now->msec - r->start_msec);
   request_time = ngx_max(request_time, 0);
-  ngx_log_stderr(0, "request time: %d %d %d", request_time, now->sec, r->start_sec);
 
   latency_record = get_latency_record(r);
 
@@ -106,10 +115,8 @@ static ngx_int_t ngx_http_record_latency(ngx_http_request_t *r)
     upstream_time += state[i].response_time;
   }
 
-  ngx_log_stderr(0, "upstream request time: %d", upstream_time);
   update_latency_record(latency_record->upstream_latency, upstream_time, main_conf);
 
-  //return ngx_http_next_header_filter(r);
   return NGX_OK;
 }
